requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dacktool==0.0.7
dbstream==0.1.28
dj-rest-auth==7.0.0
Django==5.1.2
django-allauth==65.2.0
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.23.0
google-api-python-client==1.7.11
google-auth==2.36.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.2.1
google-cloud-secret-manager==2.7.2
googleapis-common-protos==1.66.0
googleauthentication==0.0.18
grpc-google-iam-v1==0.12.7
grpcio==1.68.0
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.10
libcst==1.5.1
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.1
uritemplate==3.0.1
urllib3==2.2.3
